fix(theme): move `roadmaps` module to EE directory and update import paths
